Adjuste User Form

+++ app/Http/Controllers/Admin/UserController.php
@@ -8,6 +8,9 @@ use App\Models\Genre;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\File;
+use Illuminate\Support\Str;
+use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
@@ -18,7 +21,15 @@ class UserController extends Controller
      */
     public function index()
     {
-        $users = User::role(['Administrador', 'Empresário', 'Estagiário'])->get();
+        if (!Auth::user()->hasPermissionTo('Listar Usuários')) {
+            abort(403, 'Acesso não autorizado');
+        }
+        if (Auth::user()->hasRole('Programador')) {
+            $users = User::all();
+        } else {
+            $users = User::role(['Administrador', 'Empresário', 'Estagiário'])->get();
+        }
+
         return view('admin.users.index', compact('users'));
     }
 
@@ -29,6 +40,9 @@ class UserController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->hasPermissionTo('Criar Usuários')) {
+            abort(403, 'Acesso não autorizado');
+        }
         $genres = Genre::all();
         return view('admin.users.create', compact('genres'));
     }
@@ -41,13 +55,33 @@ class UserController extends Controller
      */
     public function store(UserRequest $request)
     {
+        if (!Auth::user()->hasPermissionTo('Criar Usuários')) {
+            abort(403, 'Acesso não autorizado');
+        }
+
         $data = $request->all();
         $data['password'] = bcrypt($request->password);
+
+        if ($request->hasFile('photo') && $request->file('photo')->isValid()) {
+            $name = Str::slug(mb_substr($data['name'], 0, 100)) . time();
+            $extenstion = $request->photo->extension();
+            $nameFile = "{$name}.{$extenstion}";
+            $data['photo'] = $nameFile;
+            $upload = $request->photo->storeAs('users', $nameFile);
+
+            if (!$upload) {
+                return redirect()
+                    ->back()
+                    ->withInput()
+                    ->with('error', 'Falha ao fazer o upload da imagem');
+            }
+        }
+
         $user = User::create($data);
 
         if ($user->save()) {
             if (!empty($request->role)) {
-                $user->syncRoles($request->roles);
+                $user->syncRoles($request->role);
             } else {
                 $user->syncRoles('Estagiário');
             }
@@ -70,12 +104,17 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        if (!Auth::user()->hasPermissionTo('Editar Usuários')) {
+            abort(403, 'Acesso não autorizado');
+        }
+
         $genres = Genre::all();
+        $roles = Role::all();
         $user = User::where('id', $id)->first();
         if (empty($user->id)) {
             abort(403, 'Acesso não autorizado');
         }
-        return view('admin.users.edit', compact('user', 'genres'));
+        return view('admin.users.edit', compact('user', 'genres', 'roles'));
     }
 
     /**
@@ -87,6 +126,10 @@ class UserController extends Controller
      */
     public function update(UserRequest $request, $id)
     {
+        if (!Auth::user()->hasPermissionTo('Editar Usuários')) {
+            abort(403, 'Acesso não autorizado');
+        }
+
         $data = $request->all();
         $user = User::where('id', $id)->first();
         if (empty($user->id)) {
@@ -97,7 +140,33 @@ class UserController extends Controller
         } else {
             $data['password'] = $user->password;
         }
+
+        if ($request->hasFile('photo') && $request->file('photo')->isValid()) {
+            $name = Str::slug(mb_substr($user->name, 0, 10)) . time();
+            $imagePath = storage_path() . '/app/public/users/' . $user->photo;
+
+            if (File::isFile($imagePath)) {
+                unlink($imagePath);
+            }
+
+            $extenstion = $request->photo->extension();
+            $nameFile = "{$name}.{$extenstion}";
+
+            $data['photo'] = $nameFile;
+
+            $upload = $request->photo->storeAs('users', $nameFile);
+
+            if (!$upload)
+                return redirect()
+                    ->back()
+                    ->withInput()
+                    ->with('error', 'Falha ao fazer o upload da imagem');
+        }
+
         if ($user->update($data)) {
+            if (!empty($request->role)) {
+                $user->syncRoles($request->role);
+            }
             return redirect()
                 ->route('admin.users.index')
                 ->with('success', 'Atualização realizada!');
@@ -117,11 +186,22 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        if (!Auth::user()->hasPermissionTo('Excluir Usuários')) {
+            abort(403, 'Acesso não autorizado');
+        }
+
         $user = User::where('id', $id)->first();
         if (empty($user->id)) {
             abort(403, 'Acesso não autorizado');
         }
+        $imagePath = storage_path() . '/app/public/users/' . $user->photo;
         if ($user->delete()) {
+            if (File::isFile($imagePath)) {
+                unlink($imagePath);
+                $user->photo = null;
+                $user->update();
+            }
+
             return redirect()
                 ->route('admin.users.index')
                 ->with('success', 'Exclusão realizada!');
